# Conflicts:
#	jmp/mp-dao/src/main/java/com/vpu/mp/dao/shop/store/StoreDao.java

+++ jmp/mp-common/src/main/java/com/vpu/mp/service/pojo/shop/config/rebate/RebateConfig.java
@@ -19,23 +19,23 @@ public class RebateConfig {
     /**
      * 商品订单分成比例
      */
-    private Double goodsOrderProportion;
+    private BigDecimal goodsOrderProportion;
     /**
      *处方药医生佣金比例
      */
-    private Double rxMedicalDoctorProportion;
+    private BigDecimal rxMedicalDoctorProportion;
     /**
      *处方药平台佣金比例
      */
-    private Double rxMedicalPlatformProportion;
+    private BigDecimal rxMedicalPlatformProportion;
     /**
      *非处方药医生佣金比例
      */
-    private Double noRxMedicalDoctorProportion;
+    private BigDecimal noRxMedicalDoctorProportion;
     /**
      *非处方药平台佣金比例
      */
-    private Double noRxMedicalPlatformProportion;
+    private BigDecimal noRxMedicalPlatformProportion;
     /**
      *医生佣金比例
      */
